[2908] styles bugfix
