+++ src/localization/locales/English.json
@@ -152,7 +152,6 @@
   "create.wallet.verify.phrase": "Verify phrase",
   "create.wallet.verify.text": "Make sure to write it down as shown. You will verify this later.",
   "create.wallet.recovery.phrase": "Recovery phrase",
-  "create.wallet.recovery.phrase.double.check": "Let’s double check",
 
   "empty.addresses": "Add an address to your watchlist and track any wallet",
   "empty.assets": "No assets yet",
@@ -173,7 +172,7 @@
   "notification.settings.transaction.alerts.switch": "Transaction alerts",
   "notification.settings.transaction.alerts.switch.text": "Get notified of transactions in your watchlist.",
 
-  "asset.your.balance": "Your balance",
+  "asset.your.balance": "Your Balance",
 
   "transaction.modal": {
     "buttons": {
@@ -253,7 +252,7 @@
   "security.confirm.passcode.description": "Re-enter your passcode to secure your wallet",
   "security.confirm.passcode.text": "Confirm passcode",
   "security.enabled": "Security Enabled!",
-  "security.enter.passcode": "Enter Passcode",
+  "security.enter.passcode": "Choose a passcode",
   "security.enter.passcode.hint": "Use a passcode you will remember. \nIt cannot be recovered.",
   "security.enter.passcode.text": "Passcode will be required for wallet access and transactions.",
   "security.enter.pin": "Enter PIN",
