+++ src/screens/CreateWallet/screens/CreateWalletSuccess/CreateWalletSuccess.styles.tsx
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+import { scale } from '@utils/scaling';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'space-between',
+    paddingHorizontal: scale(16)
+  },
+  content: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center'
+  }
+});
